deprecated go client generator
