--- bin/configs/go-experimental-go-petstore-oas2.yaml
@@ -1,6 +1,0 @@
-generatorName: go-experimental
-outputDir: samples/client/petstore/go-experimental/go-petstore
-inputSpec: modules/openapi-generator/src/test/resources/2_0/petstore-with-fake-endpoints-models-for-testing.yaml
-templateDir: modules/openapi-generator/src/main/resources/go-experimental
-additionalProperties:
-  packageName: petstore
