--- bin/configs/go-experimental-go-petstore.yaml
@@ -1,8 +1,0 @@
-generatorName: go-experimental
-outputDir: samples/openapi3/client/petstore/go-experimental/go-petstore
-inputSpec: modules/openapi-generator/src/test/resources/3_0/petstore-with-fake-endpoints-models-for-testing-with-http-signature.yaml
-templateDir: modules/openapi-generator/src/main/resources/go-experimental
-additionalProperties:
-  enumClassPrefix: "true"
-  packageName: petstore
-  disallowAdditionalPropertiesIfNotPresent: false
