--- bin/configs/go-petstore-withXml.yaml
@@ -1,8 +1,0 @@
-generatorName: go
-outputDir: samples/client/petstore/go/go-petstore-withXml
-inputSpec: modules/openapi-generator/src/test/resources/2_0/petstore-with-fake-endpoints-models-for-testing.yaml
-templateDir: modules/openapi-generator/src/main/resources/go
-additionalProperties:
-  packageName: petstore
-  withXml: "true"
-  withGoCodegenComment: "true"
