+++ bin/configs/go-petstore.yaml
@@ -1,7 +1,8 @@
 generatorName: go
 outputDir: samples/openapi3/client/petstore/go/go-petstore
-inputSpec: modules/openapi-generator/src/test/resources/3_0/petstore-with-fake-endpoints-models-for-testing.yaml
+inputSpec: modules/openapi-generator/src/test/resources/3_0/petstore-with-fake-endpoints-models-for-testing-with-http-signature.yaml
 templateDir: modules/openapi-generator/src/main/resources/go
 additionalProperties:
   enumClassPrefix: "true"
   packageName: petstore
+  disallowAdditionalPropertiesIfNotPresent: false
