+++ modules/openapi-generator/src/main/resources/go-deprecated/.travis.yml
