+++ modules/openapi-generator/src/main/resources/go-deprecated/git_push.sh.mustache
