+++ modules/openapi-generator/src/main/resources/go-deprecated/gitignore.mustache
