+++ modules/openapi-generator/src/main/resources/go-deprecated/openapi.mustache
