+++ modules/openapi-generator/src/main/resources/go-deprecated/partial_header.mustache
