+++ modules/openapi-generator/src/main/resources/go-deprecated/response.mustache
