--- modules/openapi-generator/src/main/resources/go-experimental/model.mustache
@@ -1,20 +1,0 @@
-{{>partial_header}}
-package {{packageName}}
-
-{{#models}}
-import (
-	"encoding/json"
-{{#imports}}
-	"{{import}}"
-{{/imports}}
-)
-
-{{#model}}
-{{#isEnum}}
-{{>model_enum}}
-{{/isEnum}}
-{{^isEnum}}
-{{#oneOf}}{{#-first}}{{>model_oneof}}{{/-first}}{{/oneOf}}{{^oneOf}}{{#anyOf}}{{#-first}}{{>model_anyof}}{{/-first}}{{/anyOf}}{{^anyOf}}{{>model_simple}}{{/anyOf}}{{/oneOf}}
-{{/isEnum}}
-{{/model}}
-{{/models}}
