+++ modules/openapi-generator/src/main/resources/go/model.mustache
@@ -1,43 +1,20 @@
 {{>partial_header}}
 package {{packageName}}
+
 {{#models}}
-{{#imports}}
-{{#-first}}
 import (
-{{/-first}}
+	"encoding/json"
+{{#imports}}
 	"{{import}}"
-{{#-last}}
-)
-{{/-last}}
 {{/imports}}
+)
+
 {{#model}}
 {{#isEnum}}
-// {{{classname}}} {{#description}}{{{.}}}{{/description}}{{^description}}the model '{{{classname}}}'{{/description}}
-type {{{classname}}} {{^format}}{{dataType}}{{/format}}{{#format}}{{{format}}}{{/format}}
-
-// List of {{{name}}}
-const (
-	{{#allowableValues}}
-	{{#enumVars}}
-	{{^-first}}
-	{{/-first}}
-	{{#enumClassPrefix}}{{{classname.toUpperCase}}}_{{/enumClassPrefix}}{{name}} {{{classname}}} = {{{value}}}
-	{{/enumVars}}
-	{{/allowableValues}}
-)
+{{>model_enum}}
 {{/isEnum}}
 {{^isEnum}}
-// {{classname}}{{#description}} {{{description}}}{{/description}}{{^description}} struct for {{{classname}}}{{/description}}
-type {{classname}} struct {
-{{#allVars}}
-{{^-first}}
-{{/-first}}
-{{#description}}
-	// {{{description}}}
-{{/description}}
-	{{name}} {{#isNullable}}*{{/isNullable}}{{{dataType}}} `json:"{{baseName}}{{^required}},omitempty{{/required}}"{{#withXml}} xml:"{{baseName}}{{#isXmlAttribute}},attr{{/isXmlAttribute}}"{{/withXml}}{{#vendorExtensions.x-go-custom-tag}} {{{.}}}{{/vendorExtensions.x-go-custom-tag}}`
-{{/allVars}}
-}
+{{#oneOf}}{{#-first}}{{>model_oneof}}{{/-first}}{{/oneOf}}{{^oneOf}}{{#anyOf}}{{#-first}}{{>model_anyof}}{{/-first}}{{/anyOf}}{{^anyOf}}{{>model_simple}}{{/anyOf}}{{/oneOf}}
 {{/isEnum}}
 {{/model}}
 {{/models}}
