+++ modules/openapi-generator/src/main/resources/go/model_oneof.mustache
