+++ modules/openapi-generator/src/main/resources/go/model_simple.mustache
