+++ modules/openapi-generator/src/main/resources/go/nullable_model.mustache
