+++ modules/openapi-generator/src/main/resources/go/signing.mustache
