+++ modules/openapi-generator/src/main/resources/go/utils.mustache
